fix(api): Return 200 OK on appearance deletion
The delete endpoint was configured to return a 204 No Content status,
but it was also returning a JSON response body. This violates the HTTP
specification, which states that a 204 response must not contain a
body.

This commit changes the status code to 200 OK to allow for the
successful deletion message to be returned in the response body.

Signed-off-by: WangZhiYao <393893729@qq.com>